Bugfix preventing tdm from relaunching after starting X.
Referring to
[amakovec/console-tdm/tdm:138](https://github.com/amakovec/console-tdm/blob/2f117712995647a02b55a5e4317407d476eb73cd/tdm#L138), however the line should be modified as it is below:

```bash
sid=$(dialog --stdout ${DEFOPT} --menu "TDM ${VERSION}" 0 0 0 ${prglist[@]})
```

For example, `tdm` is run in `/dev/tty1`.  When the line above is run, the command substitution above captures `stdout` in its subshell.  `X` is then started.  Once `X` is killed and `tdm` is run again in `/dev/tty1`, `stdout` is still locked to the subshell making it so that a new instance of `dialog` cannot write to it.

This bugfix makes it so that no command substitution is used, therefore
not launching a subshell, thus keeping `stdout` open for another
instance of `dialog`.