Revert sid=$?, properly fix tdm relaunch behavior.
The previous commit (e896090) changed the way that $sid was defined so
that sid=$?.  This is because it was believed that the option chosen by
the user in dialog was written to the exit status.  However, it was
discovered that the output of dialog is stored in stderr by default.
The section of code referencing dialog was rewritten for this commit so
that stderr is redirected to a temporary file tempfile=/tmp/tdm_$$.
$sid is now defined so that sid=$tempfile.  A trap has also been placed
to clean up $tempfile on exit or termination.