If X already running, just change the menu.
Fixes #26

When TDM checks for an X session already running and finds one, instead
of immediatelly falling back, continue normally but make sure that no
X session will be offered as an option in the menu.

By the time we list all of the options that will go into the menu,
if an X session was checked for and found, we do not populate the list
with xsessions.

When the time comes to show the menu to the user, if there are no options
to be shown and a running X session prevented the list from being populated
with xsessions, then instead of falling back with a message of "No sessions
found" the message will be "X started." (just like it was before this
commit).